feat: add action triggers to worker events and update related components
- Introduced `ActionTriggers` in the `workerEvent` struct to handle action commands.
- Updated `toProtoEvent` function to process `action_trigger` events and convert them to proto format.
- Enhanced `FirestoreActRunRecorder` to handle new properties in patch operations.
- Modified `startActRun` function to support recursive action triggering with a maximum depth.
- Updated `GraphCanvas` to create act nodes referencing selected nodes.
- Added a new node type `suggestion` in `GraphNodeCard` with exploration capabilities.
- Adjusted `GraphStore` to accommodate new parameters in `addQueryActNode`.
- Extended protobuf definitions to include `ActionTrigger` message and related fields.
- Updated service interfaces to handle action triggers in the RPC client.

+++ act-api/internal/adapter/firestore_act_run_recorder.go
@@ -54,8 +54,8 @@ func (r *FirestoreActRunRecorder) AppendEvent(ctx context.Context, input domain.
 				"content": op.GetContent(),
 			}
 			if op.GetSeq() > 0 {
-				payloadOp["seq"] = op.GetSeq()
-				payloadOp["expectedOffset"] = op.GetExpectedOffset()
+				payloadOp["seq"] = int64(op.GetSeq())
+				payloadOp["expectedOffset"] = int64(op.GetExpectedOffset())
 			}
 			ops = append(ops, payloadOp)
 		}
